Combine conftest with the only test file
